+++ src/Servers/Kestrel/Core/src/Internal/Http/Http1MessageBody.cs
@@ -16,29 +16,34 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http
     internal abstract class Http1MessageBody : MessageBody
     {
         protected readonly Http1Connection _context;
-        protected bool _completed;
+        private bool _readerCompleted;
 
         protected Http1MessageBody(Http1Connection context) : base(context)
         {
             _context = context;
         }
 
-        [StackTraceHidden]
-        protected void ThrowUnexpectedEndOfRequestContent()
+        public override ValueTask<ReadResult> ReadAsync(CancellationToken cancellationToken = default)
         {
-            // OnInputOrOutputCompleted() is an idempotent method that closes the connection. Sometimes
-            // input completion is observed here before the Input.OnWriterCompleted() callback is fired,
-            // so we call OnInputOrOutputCompleted() now to prevent a race in our tests where a 400
-            // response is written after observing the unexpected end of request content instead of just
-            // closing the connection without a response as expected.
-            _context.OnInputOrOutputCompleted();
+            ThrowIfReaderCompleted();
+            return ReadAsyncInternal(cancellationToken);
+        }
 
-            KestrelBadHttpRequestException.Throw(RequestRejectionReason.UnexpectedEndOfRequestContent);
+        public abstract ValueTask<ReadResult> ReadAsyncInternal(CancellationToken cancellationToken = default);
+
+        public override bool TryRead(out ReadResult readResult)
+        {
+            ThrowIfReaderCompleted();
+            return TryReadInternal(out readResult);
         }
 
         public abstract bool TryReadInternal(out ReadResult readResult);
 
-        public abstract ValueTask<ReadResult> ReadAsyncInternal(CancellationToken cancellationToken = default);
+        public override void Complete(Exception exception)
+        {
+            _readerCompleted = true;
+            _context.ReportApplicationError(exception);
+        }
 
         protected override Task OnConsumeAsync()
         {
@@ -184,12 +189,26 @@ namespace Microsoft.AspNetCore.Server.Kestrel.Core.Internal.Http
             return keepAlive ? MessageBody.ZeroContentLengthKeepAlive : MessageBody.ZeroContentLengthClose;
         }
 
-        protected void ThrowIfCompleted()
+        [StackTraceHidden]
+        private void ThrowIfReaderCompleted()
         {
-            if (_completed)
+            if (_readerCompleted)
             {
                 throw new InvalidOperationException("Reading is not allowed after the reader was completed.");
             }
+        }
+
+        [StackTraceHidden]
+        protected void ThrowUnexpectedEndOfRequestContent()
+        {
+            // OnInputOrOutputCompleted() is an idempotent method that closes the connection. Sometimes
+            // input completion is observed here before the Input.OnWriterCompleted() callback is fired,
+            // so we call OnInputOrOutputCompleted() now to prevent a race in our tests where a 400
+            // response is written after observing the unexpected end of request content instead of just
+            // closing the connection without a response as expected.
+            _context.OnInputOrOutputCompleted();
+
+            KestrelBadHttpRequestException.Throw(RequestRejectionReason.UnexpectedEndOfRequestContent);
         }
     }
 }
